% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/afaces.R
\name{as.afaces}
\alias{afaces}
\alias{as.afaces}
\title{Convert an object to an afaces object}
\usage{
as.afaces(x)

afaces(x)
}
\arguments{
\item{x}{an object that should be an array}
}
\description{
Simply adds 'afaces' to the object, if not already set. No more check so far.
}
\examples{
a <- array(1:27, dim=rep(3, 3))
a
class(a)
a2 <- as.afaces(a) # which is equivalent to afaces(a)
a2
class(as.afaces(a2))
}

