% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/afaces.R
\name{as.matrix.afaces}
\alias{as.matrix.afaces}
\title{Converts an afaces object to a wide matrix}
\usage{
\method{as.matrix}{afaces}(x, colnames = FALSE, ...)
}
\arguments{
\item{x}{afaces object}

\item{colnames}{logical, whether to retain rows/cols names of the afaces object}

\item{...}{only used to maintain generic as.matrix}
}
\description{
If the array comes as an object with (i, j, k) pixels, returns
a matrix of (k, i*j) pixels. Basically takes all the pixels from
top to bottom and left to right, convert them as a vector and turns them into
rows of the wide matrix.
}
\examples{
a <- array(1:27, dim=rep(3, 3))
a <- afaces(a)
a[,,1]
as.matrix(a)
}

