% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{avg_face}
\alias{avg_face}
\title{Calculate Average Face}
\usage{
avg_face(td)
}
\arguments{
\item{td}{List of arrays. Training data.}
}
\value{
Returns average face as array.
}
\description{
Calculates average data (here: face) of the given dataset td.
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
# Get average face
avg_face(td)
}
