% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ef.R
\name{classification_ef}
\alias{classification_ef}
\title{Find the best matching faces of an 'imageset_ef' for a given 'image_ef'}
\usage{
classification_ef(img, td, nclosest = 3, neigenfaces = 15, quick = FALSE)
}
\arguments{
\item{img}{an object of class 'image_ef', the functions searches for similar images to this image}

\item{td}{an object of class 'imageset_ef', set of images where the function searches for similar images}

\item{nclosest}{(optional, default: \code{nclosest=3}) an numeric vector of length 1, number of similar images to be ruturned}

\item{neigenfaces}{(optional, default: \code{neigenfaces=15}) an numeric vector of length 1, number of eigenfaces used for finding the closest faces}

\item{quick}{logical vector of length  (TRUE or FALSE), see \code{PCA}}
}
\value{
'imageset_ef' object containing the \code{nclosest} images from \code{td}, which are 'closest' to \code{img}
}
\description{
Finds the closest \code{nclosest} images in an object of class 'imageset_ef'. For the determination
of the closest images \code{neigenfaces} are used. There are two possible modes (\code{quick = TRUE}, \code{quick = FALSE}).
For further explanation see \code{?PCA}.
}
\examples{
# Import Olivetti-faces
td <- load_imageset_ef("../olivetti_X.csv", c(64,64))
img <- td[[1]]
closest <- classification_ef(img, td)

}
