% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eigenfaces_package.R
\docType{data}
\name{pain}
\alias{pain}
\title{Pain expression subset from Stirling (PICS)}
\format{84 face images, fixed eye location, 7 expressions (not sad) from each of 12 women.
Resolution: 181x241 monochrome
\describe{an afaces object}}
\source{
"Pain expression subset" found there \url{http://pics.psych.stir.ac.uk/2D_face_sets.htm}
}
\description{
Converted as an afaces object
}
\keyword{Datasets}

