% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenfaces-package.R
\docType{package}
\name{eigenfaces-package}
\alias{eigenfaces}
\alias{eigenfaces-package}
\title{eigenfaces: PCA on Eigenfaces}
\description{
Performs PCA and multiple sidetasks on the Olivetti-Dataset or 
    similar datasets.
}
\section{}{
NA
}

\references{
\url{https://www.bytefish.de/pdf/eigenfaces.pdf}
}
\keyword{internal}
