% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpaceProjection.R
\name{FSP.image_ef}
\alias{FSP.image_ef}
\title{Projection of an 'image_ef' object onto a set of orthogonal eigenfaces}
\usage{
\method{FSP}{image_ef}(obj, eigenfaces, avgFace, showCoefficients = FALSE, ...)
}
\arguments{
\item{obj}{an object of class 'image_ef'}

\item{eigenfaces}{an object of class 'imageset_ef', eigenfaces of an 'imageset_ef' object td}

\item{avgFace}{an object of class 'image_ef', average face of td}

\item{showCoefficients}{(optional) logical vector of length 1 (TRUE or FALSE)}

\item{...}{additional arguments}
}
\value{
case \code{showCoefficients=FALSE}: a list of length 1 containing an object of class 'image_ef',
projection of \code{img} onto the linear span of \code{eigenfaces}
case \code{showCoefficients=TRUE}: a list of length 2 additionally containing a numeric vector where the
coefficients for the representation with the eigenfaces are saved
}
\description{
The projection of the vector onto the linear span of the eigenfaces is computed
by projecting an 'image_ef' onto every single vector of the set of eigenfaces. A given average
face is substracted before the projection and added onto the projection afterwards.
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
img <- td[[42]]
eigenfaces <- get_eigenfaces(td, nfaces = 50, quick=TRUE)
avgFace <- avg_face(td)
projection <- FSP(img, eigenfaces, avgFace, showCoefficients=FALSE)[[1]]
}
