% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpaceProjection.R
\name{FSP.imageset_ef}
\alias{FSP.imageset_ef}
\title{Projection of an 'imageset_ef' object onto a set of orthogonal eigenfaces}
\usage{
\method{FSP}{imageset_ef}(td, eigenfaces, avgFace, showCoefficients = FALSE)
}
\arguments{
\item{td}{an object of class 'imageset_ef'}

\item{eigenfaces}{an object of class 'imageset_ef', eigenfaces of an 'imageset_ef' object td}

\item{avgFace}{an object of class 'image_ef', average face of td}

\item{showCoefficients}{(optional) logical vector of length 1 (TRUE or FALSE)}
}
\value{
case \code{showCoefficients=FALSE}: a list of length 1 containing a list of objects of class 'image_ef',
projections of the elements of \code{td} onto the linear span of \code{eigenfaces}
case \code{showCoefficients=TRUE}: a list of length 2 additionally containing a list of numeric vectors where the
coefficients for the representation with the eigenfaces are saved.
}
\description{
applies \code{FSP.image_ef} to every single 'image_ef' object of the 'imageset_ef' object.
The projected faces and the coefficients are returned as a list each (combined to another list)
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
eigenfaces <- get_eigenfaces(td, nfaces = 50)
avgFace <- avg_face(td)
projection <- FSP(td, eigenfaces, avgFace, showCoefficients=FALSE)[[1]]
}
