% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{get_eigenfaces}
\alias{get_eigenfaces}
\title{Calculation of the eigenvectors and eigenvalues of the covariance matrix}
\usage{
get_eigenfaces(td, nfaces = 15, quick = FALSE)
}
\arguments{
\item{td}{List of arrays. Training data.}

\item{nfaces}{Amount of Eigenfaces}
}
\value{
Returns n=1,...,nfaces Eigenfaces.
}
\description{
Makes use of the PCA function to perform the principle component analysis
}
\examples{
# Import Olivetti-faces
td <- load_imageset_ef("olivetti_X.csv", c(64,64))
get_eigenfaces(td, 9)
}
