% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{image_ef}
\alias{image_ef}
\title{Object image_ef}
\usage{
image_ef(x)
is.image_ef(x)
}
\arguments{
\item{x}{List of arrays.}
}
\value{
Returns image in grayscale.  ##Stimmt das?
}
\description{
Creates objects of class image_ef.
They are printed as images in grayscale.
is.image_ef is more a general test of an object being interpretable as an image.
}
\examples{
##Todo: bessere Beispiele
image_ef(x)
is.image_ef(x)
}
