% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{imageset_ef}
\alias{imageset_ef}
\title{Creates an Object of class 'imageset_ef'}
\usage{
imageset_ef(lst)
}
\arguments{
\item{lst}{list of objects capable of being used as input for \code{image_ef()} function.}
}
\value{
object of class 'imageset_ef',a list consisting of objects of class 'image_ef'.
}
\description{
Creates an Object of class 'imageset_ef'
}
\examples{
img1 <- image_ef(matrix(c(1,0,0,1), nrow=2))
img2 <- image_ef(matrix(c(0,1,1,0), nrow=2))
lst <- list(img1, img2)
td <- imageset_ef(lst)
}
