% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import.R
\name{import_faces}
\alias{import_faces}
\title{Import a set of .jpg images to an afaces object}
\usage{
import_faces(path = getwd())
}
\arguments{
\item{path}{the path to .jpg (working directory by default)}
}
\description{
Import a set of .jpg images to an afaces object
}

