% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{is.image_ef}
\alias{is.image_ef}
\title{Test if the input is of class 'image_ef'}
\usage{
is.image_ef(img)
}
\arguments{
\item{img}{An object to be tested.}
}
\value{
TRUE, if \code{img} is of class 'image_ef', FALSE otherwise
}
\description{
Test if the input is of class 'image_ef'
}
\examples{
img <- image_ef(c(1,2,3,4))
is.image_ef(img)

noImg <- "a"
is.image_ef(noImg)
}
