% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{is.imageset_ef}
\alias{is.imageset_ef}
\title{Tests if an object is of class 'imageset_ef'}
\usage{
is.imageset_ef(td)
}
\arguments{
\item{td}{An object to be tested}
}
\value{
logical vector of legnth 1, (TRUE or FALSE)
}
\description{
Tests if an object is of class 'imageset_ef'
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
# Test
is.imageset_ef(td)

is.imageset_ef(c(1,2,3,4))
}
