% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import.R
\name{lf_structure}
\alias{lf_structure}
\title{Extracts grouping structure from filenames}
\usage{
lf_structure(lf, ...)
}
\arguments{
\item{lf}{a list of files}

\item{...}{useless, used for genericity.}
}
\description{
Can be called either on an object returned by \link{list.files} or directly
on an afaces object. In the latter case, \code{dimanmes(afaces)[3]} is used.
See Momocs' lf_structure.
}

