% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_and_save.R
\name{load_any_imageset}
\alias{load_any_imageset}
\title{Load any dataset}
\usage{
load_any_imageset(folderpath, filetype)
}
\arguments{
\item{folderpath}{str, relative path to the folder (incl. the foldername)}

\item{filetype}{str, lower letter filetype}
}
\value{
td, List of arrays. Training Data.
}
\description{
Makes it possible to not only load .csv files (as in the case of the Olivetti
Dataset) but also to load seperate .png, .jpeg or .bmp files as an imageset.
}
\examples{
td <- load_any_imageset("inst/extdata/testfiles", "jpg")
imageShow_ef(td[[1]])
}
