% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{load_classes_ef}
\alias{load_classes_ef}
\title{Load csv-File and Create Array}
\usage{
load_classes_ef(path, header = FALSE)
}
\arguments{
\item{path}{Path to dataset. Needs to be in csv-format.}
}
\value{
Returns 1D-array.
}
\description{
Loads csv-file and creates an array containing the respective class labels of
the dataset.
}
\examples{
# Load classes of Olivetti-Dataset
classes <- load_classes_ef("olivetti_y.csv")
}
