% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{load_imageset_ef}
\alias{load_imageset_ef}
\title{Load csv-File and Create Imageset_ef}
\usage{
load_imageset_ef(path, imgDim)
}
\arguments{
\item{path}{Path to dataset. Needs to be in csv-format.}

\item{imgDim}{Tupel. Two dimensional vector to indicate dimension of seperate
images.}
}
\value{
Returns dataset as \code{imageset_ef} object.
}
\description{
Loads csv-file and saves it as imageset_ef. Imageset_ef will be a list of
multiple imgDim x imgDim arrays.
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
}
