% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eigenfaces_package.R
\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{eigenfaces}
\description{
Eigenfaces are mainly used for face recognition and face morphometrics,
using dimensionality reduction via eigenvectors (principal component analysis)
on a set of faces, coming as .jpg images.
}
\details{
In many aspects, it is a plugin of the Momocs package: it both takes profits and extends it.
To cite eigenfaces in publications: \code{citation('eigenfaces')}. (Paper in prep.)
}

