% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{normalize.image_ef}
\alias{normalize.image_ef}
\title{Normalize of image_ef objects}
\usage{
\method{normalize}{image_ef}(img)
}
\arguments{
\item{img}{an object of class 'image_ef'.}
}
\value{
Returns normalized version of img.
}
\description{
Normalize an image_ef object.
}
\details{
The image is subtracted by the sum of all images divided by their length.
}
\examples{
##Todo: bessere Beispiele
normalize(img)
}
