% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{normalize.imageset_ef}
\alias{normalize.imageset_ef}
\title{Normalization of an imageset_ef}
\usage{
\method{normalize}{imageset_ef}(td)
}
\arguments{
\item{td}{List of arrays. Training data.}
}
\value{
Returns normalized version of td.
}
\description{
Normalizes all entries in the imageset_ef list.
}
\examples{
# Import Olivetti-faces
td <- load_imageset_ef("olivetti_X.csv", c(64,64))
# Normalize
normalize(td)
}
