% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions_ef.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(obj)
}
\arguments{
\item{obj}{an object to be normalized (e.g. of class 'image_ef', 'imageset_ef')}
}
\description{
Normalization
}
