% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olivetti_X_data.R
\docType{data}
\name{olivetti_X}
\alias{olivetti_X}
\title{Olivetti Dataset}
\format{
A .csv file
}
\source{
\url{https://scikit-learn.org/0.19/modules/generated/sklearn.datasets.fetch_olivetti_faces.html#sklearn.datasets.fetch_olivetti_faces}
}
\description{
The Olivetti Dataset contains 40 Persons which are photographed at 10 different times,
variying lighting facial express and facial detail.
So the Dataset contains 400 face images.
All faces have black background and are grey level. The size is 64*64.
The images were taken between April 1992 and April 1994.
The Image pixel values are scaled to [0, 1] interval.
The names of the persons are encoded to an integer of 0 to 39.
}
\keyword{datasets}
