% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/afaces.R
\name{PCA.afaces}
\alias{PCA.afaces}
\title{Principal component analysis on afaces objects}
\usage{
\method{PCA}{afaces}(x, scale. = TRUE, center = TRUE, fac = data.frame())
}
\arguments{
\item{x}{an afaces object}

\item{scale.}{logical, whether to scale the matrix}

\item{center}{logical, whether to center the matrix}

\item{fac}{data.frame (optionnal) that can be use with Momocs' PCA facilities.}
}
\value{
a list with the same components as \link{prcomp} plus
\code{$scale}, \code{$center}, \code{$mshape} and \code{$dim}.
}
\description{
Takes an afaces objet, uses as.matrix.afaces to convert it into a wide matrix,
then call \link{prcomp} to perform the PCA. Returns a \code{PCA} object ala Momocs
on which \code{plot}, \code{MANOVA}, etc. are defined.
}

