% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pfaces.R
\name{PCA2face}
\alias{PCA2face}
\title{Given a pfaces object, reconstruct faces}
\usage{
PCA2face(PCA, pos = c(0, 0), amp = 1, xax = 1, yax = 2)
}
\arguments{
\item{PCA}{a pfaces object}

\item{pos}{a vector of length 2 or a 2-col matrix specifying the position
on the factorail place (xax, yax) to use to reconstruct faces
(\code{(c(0, 0)} by default)}

\item{amp}{numeric an amplification factor (1 by default)}

\item{xax}{integer which PC to use as the 1st axis (1st by default)}

\item{yax}{integer which PC to use as the 2nd axis (2nd by default)}
}
\description{
Given a pfaces object, reconstruct faces and returns an afaces object.
}

