% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/afaces.R
\name{plot.afaces}
\alias{plot.afaces}
\title{Plot method for afaces objects}
\usage{
\method{plot}{afaces}(x, y = 1, ...)
}
\arguments{
\item{x}{afaces object}

\item{y}{the id of the face to plot (i-th slice of the array). 1 by default}

\item{...}{only used to maintain generic plot}
}
\description{
Plots a particular face from an afaces object
}
\examples{
a <- array(runif(27), dim=rep(3, 3))
a <- afaces(a)
plot(a, 1) # not very sexy ; same as plot(a)
plot(a, 2)
}

