% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userfriendly_functions.R
\name{plot_olivetti}
\alias{plot_olivetti}
\title{Plot of Olivetti Dataset}
\usage{
plot_olivetti(Persons = 1:10, number = 10)
}
\arguments{
\item{Persons}{desired persons of the Olivetti Dataset. It can be an integer or double vector or a double value. Default gives first 10 different persons.}

\item{number}{desired number of photos of the selected persons. Default is 10.}
}
\value{
Returns Olivetti-faces from the Olivetti Dataset.
}
\description{
Plot different face images of the Olivetti Dataset.
}
\details{
The Olivetti Dataset contains 40 Persons which are photographed at 10 different times, variying lighting facial express and facial detail.
All faces have black background and are grey level. The size is 64*64. The images were taken between April 1992 and April 1994.
}
\examples{
plot_olivetti()
plot_olivetti(2,2)
plot_olivetti(1:3,10)
plot_olivetti(c(1,2),10)
}
