% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{print.image_ef}
\alias{print.image_ef}
\title{Show an image_ef object as image in the plots section}
\usage{
\method{print}{image_ef}(img, ...)
}
\arguments{
\item{img}{An image_ef object.}

\item{...}{TODO}
}
\description{
An implementation of the \code{print} function for 'image_ef' objects.
}
\examples{
img <- image_ef(matrix(c(1,0,1,0), nrow=2))
}
