% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpaceProjection.R
\name{proj}
\alias{proj}
\title{Projection of a vector onto a unit vector}
\usage{
proj(x, y)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{y}{numeric vector or matrix}
}
\value{
scalar product of x with unit vector in y-direction
}
\description{
Projection of a vector onto a unit vector
}
\examples{
proj(c(1,1,1), c(1,0,0))
}
