% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny_ef}
\alias{runShiny_ef}
\title{Run Shiny App for Olivetti Dataset}
\usage{
runShiny_ef()
}
\value{
Command to open shiny app.
}
\description{
Opens customized shiny app for the Olivetti Dataset. No need to load Dataset
before usage. Number of eigenfaces is limited to 400.
Functions: 1. Display any image of the dataset. 2. Reconstruct this image
with an amount of eigenfaces of your choice. 3. Display the first 400
eigenfaces.
}
\examples{
# Opens Shiny App for Olivetti Dataset
runShiny_ef()
}
