% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny_general}
\alias{runShiny_general}
\title{Run Shiny App for Any Dataset}
\usage{
runShiny_general(td)
}
\arguments{
\item{td}{Object of class imageset_ef. Training data.}
}
\value{
Command to open shiny app.
}
\description{
Opens shiny app for any suitable dataset. Number of eigenvectors is limited to
400.
Functions: 1. Display any image of the dataset. 2. Reconstruct this image
with an amount of eigenvectors of your choice. 3. Display the first 400
eigenvectors.
}
\examples{
# Opens Shiny App for (e.g.) MNIST Dataset
td <- load_imageset_ef("inst/extdata/mnist_X.csv", c(28,28))
runShiny_general(td)
}
