% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_file_read_in.R
\name{save_image_ef}
\alias{save_image_ef}
\title{Save image}
\usage{
save_image_ef(im, filename, filetype, path = default)
}
\arguments{
\item{im}{arr, array-like structure}

\item{filename}{str, working names}

\item{filetype}{str, "jpg" or "png"}

\item{path}{str, optional. Default is current wd.}
}
\value{

}
\description{
Save any suitable array as image.
}
\examples{
td <- load_any_imageset("testfiles", "jpg")
save_image_ef(td[[3]], "test_save", "jpg", path = "testfiles")
}
