% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/domestic.R
\name{shave}
\alias{shave}
\title{Shaves a matrix}
\usage{
shave(m, max = 1, min = 0)
}
\arguments{
\item{m}{matrix to shave}

\item{max}{the ceiling}

\item{min}{the floor}
}
\description{
Ie, everything above max is set to max; everything below min is set to min.
Used to avoid grey levels below 0 or above 1 when reconstructing faces,
and consequently avoid errors with plot method.
That explains white or dark areas on extreme PC ranges.
}

