% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ef.R
\name{distance_ef}
\alias{distance_ef}
\title{Euclidean distance between the coefficient vectors of the projection of two images}
\usage{
distance_ef(coeffs1, coeffs2)
}
\arguments{
\item{coeffs1}{numeric vector, first coefficient vector}

\item{coeffs2}{numeric vector, second coefficient vector}
}
\value{
numeric vector of length 1, the euclidean distance of \code{coeffs1} and \code{coeffs2}
}
\description{
Euclidean distance between the coefficient vectors of the projection of two images
}
\examples{
# Import Olivetti-faces
td <- load_imageset_ef("olivetti_X.csv", c(64,64))
#compute coefficients
eigenfaces <- get_eigenfaces(td, nfaces = 50)
avgFace <- avg_face(td)
coeffs1 <- FSP(td[[1]], eigenfaces, avgFace, showCoefficients=TRUE)[[2]]
coeffs2 <- FSP(td[[2]], eigenfaces, avgFace, showCoefficients=TRUE)[[2]]
distance_ef(coeffs1, coeffs2)
}
