% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenfaces-package.R
\docType{package}
\name{eigenfaces-package}
\alias{eigenfaces}
\alias{eigenfaces-package}
\title{Eigenface package}
\description{
Performs PCA and multiple sidetasks on the Olivetti-Dataset or 
    similar datasets.
}
\details{
The package contains the principal component analyses of a Dataset of face images (here: Olivetti Data).
         It provides several functionalities:

         * output of the n important Eigenfaces of a Dataset
         * Finding of similar faces for a given new face by projection on the main components.
         * Presentation of figures from the Dataset with reduced number of main components.
}
\references{
\url{https://www.bytefish.de/pdf/eigenfaces.pdf}
}
\author{
\strong{Maintainer}: Aaron Osburg \email{Osburg@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Marcel Fischer \email{Marcel.Fischer@stud.uni-heidelberg.de}
  \item Ida Jandl \email{Jandl@stud.uni-heidelberg.de}
}

}
\keyword{internal}
