% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenfaces-package.R
\docType{package}
\name{eigenfaces-package}
\alias{eigenfaces}
\alias{eigenfaces-package}
\title{eigenfaces: PCA on Eigenfaces}
\description{
Performs PCA and multiple sidetasks on the Olivetti-Dataset or 
    similar datasets.
}
\section{}{
NA
}

\references{
\url{https://www.bytefish.de/pdf/eigenfaces.pdf}
}
\author{
\strong{Maintainer}: Aaron Osburg \email{Osburg@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Marcel Fischer \email{Marcel.Fischer@stud.uni-heidelberg.de}
  \item Ida Jandl \email{Jandl@stud.uni-heidelberg.de}
}

}
\keyword{internal}
