% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions_ef.R
\name{FSP}
\alias{FSP}
\title{Feature Space Projection}
\usage{
FSP(obj, eigenfaces, avgFace, showCoefficients = TRUE)
}
\description{
Projects an image or set of images into feature space defined by a set of eigenfaces
}
\details{
For details see FSP.image_ef and FSP.imageset_ef
}
