% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{get_eigenfaces}
\alias{get_eigenfaces}
\title{Calculate the eigenvectors and eigenvalues of the covariance matrix}
\usage{
get_eigenfaces(td, nfaces = 15, quick = FALSE)
}
\arguments{
\item{td}{an object of class 'imageset_ef'. List of arrays. Training data.}

\item{nfaces}{a numeric vector of length 1. The desired number of eigenfaces.}

\item{quick}{a logical vector of length 1. Determines if the PCA is done with 'quick=TRUE' or 'quick=FALSE'.}
}
\value{
Returns n=1,...,nfaces Eigenfaces.
}
\description{
Makes use of the PCA function to perform the principle component analysis.
The data is normalized before performing the PCA.
}
\examples{
# Import Olivetti-faces
olivetti <- system.file("extdata","olivetti_X.csv",package="eigenfaces")
td <- load_imageset_ef(olivetti, c(64,64))
get_eigenfaces(td, 9)
PCA(td,showEigenvals=FALSE, quick=TRUE)
}
\references{
Marinovsky F., Wagner P., Gesichtserkennung mit Eigenfaces, FH Zittau/Görlitz
}
