% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{image_ef}
\alias{image_ef}
\title{Create an image_ef object from x.}
\usage{
image_ef(x)
}
\arguments{
\item{x}{A matrix or an atomic vector of type numeric.}
}
\value{
An object of class \code{image_ef} and attribute \code{dim}. Values are taken from x.
the value of \code{dim} corresponds to the value of \code{dim(x)}. If
\code{dim} is not set for \code{x}, \code{c(length(x), 1)} is used instead.
}
\description{
Create an image_ef object from x.
}
\examples{
image_ef(matrix(c(1,2,3,4), nrow=2))
image_ef(c(1,2,3,4))
}
