% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ef.R
\name{normalize.image_ef}
\alias{normalize.image_ef}
\title{Normalization of an object of class 'image_ef'}
\usage{
\method{normalize}{image_ef}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'image_ef'.}

\item{...}{additional arguments}
}
\value{
An object of class 'image_ef'. The pixel values are those of img minus the mean over all pixels of img.
}
\description{
Subtracts the mean over all pixel from every single pixel and returns the result.
}
\references{
https://www.bytefish.de/pdf/eigenfaces.pdf
}
