% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_and_save.R
\name{save_image_ef}
\alias{save_image_ef}
\title{Save image}
\usage{
save_image_ef(im, filename, filetype, path = NULL)
}
\arguments{
\item{im}{arr, array-like structure}

\item{filename}{str, working names}

\item{filetype}{str, "jpg" or "png"}

\item{path}{str, optional. Default is current wd.}
}
\value{

}
\description{
Save any suitable array as image.
}
\examples{
folder <- system.file("extdata","testfiles",package="eigenfaces")
td <- load_any_imageset(folder, "jpg")
save_image_ef(td[[3]], "test_save", "jpg", path = folder)
}
