% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageset_ef.R
\name{subtract_avg_face}
\alias{subtract_avg_face}
\title{Subtract Average Face}
\usage{
subtract_avg_face(td)
}
\arguments{
\item{td}{List of arrays. Training data.}
}
\value{
Returns td - average face.
}
\description{
Makes use of avg_face() function to subtract the average data (here: face)
of a given data set.
}
